third_party/WebKit/LayoutTests/platform/win/fast/forms/selection-setSelectionRange-frameselection-expected.txt

This is a testharness.js-based test.
PASS setSelectionRange() should not update FrameSelection if the target element is not focused. 
PASS setRangeText() should not update FrameSelection if the target element is not focused. 
Harness: the test ran to completion.
